Making fw take floats
git-svn-id: http://papert.googlecode.com/svn/trunk@74 e7052fda-7f78-11dd-a143-0dd59a09f4aa

+++ logo.js
@@ -78,12 +78,12 @@ Logo.prototype.setup = function () {
 
     this.addCommand('forward',1,['fw','fd'], function (a) { 
         if (parseFloat(a[0]) != a[0]) return new Token('error','Can only go forward a number, not '+a[0])
-        this.turtle.forward(parseInt(a[0]));
+        this.turtle.forward(parseFloat(a[0]));
     });
     
     this.addCommand('backward',1,['bw','bk','back'], function (a) { 
         if (parseFloat(a[0]) != a[0]) return new Token('error','Can only go backward a number, not '+a[0])
-        this.turtle.backward(parseInt(a[0]));
+        this.turtle.backward(parseFloat(a[0]));
     });
     this.addCommand('right',1,['rt'], function (a) { 
         if (parseFloat(a[0]) != a[0]) return new Token('error','Can only turn right a number of degrees, not '+a[0])
